+++ src/main/java/net/soulsweaponry/client/hud/EffectHudOverlay.java
@@ -11,6 +11,8 @@ import net.minecraft.util.Identifier;
 public abstract class EffectHudOverlay implements HudRenderCallback {
 
     private int yOffset = 0;
+    public static final int BAR_WIDTH = 182;
+    public static final int BAR_HEIGHT = 5;
 
     @Override
     public void onHudRender(DrawContext drawContext, float v) {
@@ -29,8 +31,8 @@ public abstract class EffectHudOverlay implements HudRenderCallback {
                     RenderSystem.setShaderColor(1f, 1f, 1f, 1f);
 
                     drawContext.drawTexture(this.getTexture(), barX - 25, barY - 10, 0, 0, 25, 25, 207 ,25); // Icon
-                    drawContext.drawTexture(this.getTexture(), barX, barY, 25, 10, 182, 5, 207, 25); // Empty
-                    drawContext.drawTexture(this.getTexture(), barX, barY, 25, 15, pixelOffset, 5, 207, 25); // Filled
+                    drawContext.drawTexture(this.getTexture(), barX, barY, 25, 10, BAR_WIDTH, BAR_HEIGHT, 207, 25); // Empty
+                    drawContext.drawTexture(this.getTexture(), barX, barY, 25, 15, pixelOffset, BAR_HEIGHT, 207, 25); // Filled
                 }
             }
         }